+++ packages/frontend/src/pages/index.ts
@@ -9,7 +9,7 @@ import { getMultisigReportDownloadPage } from './multisig-report'
 import { outputPages } from './output'
 import { Page, PagesData } from './Page'
 import { getActivityPage } from './scaling-activity'
-import { getL2AssetsPage } from './scaling-l2assets'
+import { getL2AssetsPage } from './scaling-detailedTvl'
 import { getProjectPages } from './scaling-projects'
 import { getRiskPage } from './scaling-risk'
 import { getTvlPage } from './scaling-tvl'
