+++ packages/frontend/src/pages/scaling-activity/props/getProps.ts
@@ -31,7 +31,7 @@ export function getProps(
       ),
       footer: getFooterProps(config),
       showActivity: config.features.activity,
-      showL2Assets: config.features.l2assets,
+      showL2Assets: config.features.detailedTvl,
       milestones: config.features.milestones ? config.milestones : [],
     },
     wrapper: {
