+++ packages/frontend/src/pages/scaling-detailedTvl/props/getDetailedTvlView.ts
