+++ packages/frontend/src/pages/scaling-detailedTvl/props/getPageMetadata.ts
@@ -5,7 +5,7 @@ export function getPageMetadata(): PageMetadata {
     title: 'L2BEAT – The state of the layer two ecosystem',
     description:
       'L2BEAT is an analytics and research website about Ethereum layer 2 scaling. Here you will find in depth comparison of major protocols live on Ethereum today.',
-    image: 'https://l2beat.com/meta-images/overview-scaling-l2assets.png',
+    image: 'https://l2beat.com/meta-images/overview-scaling-detailedTvl.png',
     url: 'https://l2beat.com/',
   }
 }
