+++ packages/frontend/src/pages/scaling-detailedTvl/props/index.ts
