+++ packages/frontend/src/pages/scaling-detailedTvl/view/DetailedTvlView.tsx
