+++ packages/frontend/src/pages/scaling-detailedTvl/view/L2AssetsPage.tsx
