+++ packages/frontend/src/pages/scaling-tvl/props/getProps.ts
@@ -37,7 +37,7 @@ export function getProps(
       ),
       footer: getFooterProps(config),
       showActivity: config.features.activity,
-      showL2Assets: config.features.l2assets,
+      showL2Assets: config.features.detailedTvl,
       showMultisigReport: config.features.multisigReport,
       milestones: config.features.milestones ? config.milestones : [],
     },
