+++ packages/frontend/src/utils/project/getChart.ts
@@ -21,7 +21,7 @@ export function getChart(
     hasActivity:
       config?.features.activity &&
       !!activityApiResponse?.projects[project.id.toString()],
-    hasDetailedTvl: config?.features.l2assets,
+    hasDetailedTvl: config?.features.detailedTvl,
     milestones: config?.features.milestones ? project.milestones : [],
     isUpcoming: project.isUpcoming ?? project.config.escrows.length === 0,
   }
